pyproject.toml
[tool.poetry]
name = "codegraphclassification"
version = "0.1.0"
description = ""
authors = ["Cezar Sas <cezar.sas@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
igraph = "^0.9.11"
hydra-core = "^1.2.0"
more-itertools = "^8.13.0"
hydra-colorlog = "^1.2.0"
fasttext = "^0.9.2"
spacy = "^3.3.1"
gensim = "^4.2.0"
pandas = "^1.4.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
